Project Infrastructure: Makefile addition to project.
Adding a Makefile to the project.

Current setup will have the Makefile create the build directory
structure to have the build command output the executables to.

Have a standard build option as well as a build option for debugging
symbols.

The Makefile assumes that the implementation and include files are
stored in the src and include directories respectfully.