Modification of the Makefile.
Made the default make target print out the targets of the makefile.
Works by grepping through the makefile for '#-' while ignoring lines
with grep in it.

The lines preceding the targets have the '#-' at the start, followed by
the usage of that target.

Added in a build target that simple compiles the project.